separated main function from CLI
