+++ conf/__init__.py
@@ -1,0 +1,1 @@
+"""Configuration files in BTK (this init file is required by hydra)."""
